+++ config/ragnarok.php
@@ -3,5 +3,6 @@
 return [
     'sinks' => [
         \TromsFylkestrafikk\RagnarokConsat\Sinks\SinkConsat::class,
+        \TromsFylkestrafikk\RagnarokRuter\Sinks\SinkRuter::class,
     ],
 ];